Arreglando merge conflicts por cambio de nombre de archivos
